Wifi: Reduce the frequency of scans in disconnected state and display ON
When STA is not connected, display is ON and there are no saved profiles
then reduce frequent scans due to framework disconnected scan by increasing
the time mDisconnectedScanPeriodMs to 5 minutes, since there is already a
periodic scan for every mSupplicantScanIntervalMs(15 seconds). If there
are saved profiles then keep mDisconnectedScanPeriodMs to 10 seconds.

Change-Id: Ifcf17d1403f8df2d7c031a2f04a6376f5c4c4366
CRs-Fixed: 822066