Wifi: Screen state to set/reset mEnableBackgroundScan
mEnableBackgroundScan is set to true only when the screen state change is
handled from the disconnected state. Thus, the screen state change during
the connected state shall not result in mEnableBackgroundScan set to true.
This is required for triggering PNO once the Wi-Fi state machine transits from
connected to disconnected state when the screen is OFF.

This commit ties mEnableBackgroundScan value to the screen state.

Change-Id: Icf30c024d7167570cc8b150c1d73a2400bf7a210
CRs-Fixed: 810478